pyproject.toml
[project]
name = "vibehub-testing"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth>=2.43.0",
    "google-auth-oauthlib>=1.2.2",
    "python-dotenv>=1.2.1",
]
